+++ app/src/main/java/io/github/lazyimmortal/sesame/model/task/antOcean/AntOcean.java
@@ -14,6 +14,7 @@ import io.github.lazyimmortal.sesame.entity.AlipayAntOceanAntiepTaskList;
 import io.github.lazyimmortal.sesame.entity.AlipayUser;
 import io.github.lazyimmortal.sesame.hook.ApplicationHook;
 import io.github.lazyimmortal.sesame.model.base.TaskCommon;
+import io.github.lazyimmortal.sesame.model.normal.base.BaseModel;
 import io.github.lazyimmortal.sesame.model.task.antFarm.AntFarm.TaskStatus;
 import io.github.lazyimmortal.sesame.model.task.antForest.AntForestRpcCall;
 import io.github.lazyimmortal.sesame.util.Log;
@@ -284,6 +285,7 @@ public class AntOcean extends ModelTask {
                         }
                     }
                 }
+                BaseModel.sleepTaskInterval();
             }
         }
         catch (Throwable t) {
@@ -301,6 +303,7 @@ public class AntOcean extends ModelTask {
                     checkReward(cleanRewardVOS);
                     Log.forest("神奇海洋🐳清理[" + UserIdMap.getMaskName(userId) + "]海域");
                 }
+                BaseModel.sleepTaskInterval();
             }
         }
         catch (Throwable t) {
@@ -428,6 +431,7 @@ public class AntOcean extends ModelTask {
                 if (MessageUtil.checkResultCode(TAG, jo)) {
                     Log.forest("神奇海洋🐳[学习海洋科普知识]#获得[潘多拉能量*1]");
                 }
+                BaseModel.sleepTaskInterval();
             }
         }
         catch (Throwable t) {
@@ -467,6 +471,7 @@ public class AntOcean extends ModelTask {
                 JSONObject bizInfo = new JSONObject(jo.getString("bizInfo"));
                 String taskTitle = bizInfo.getString("taskTitle");
                 receiveReplicaTaskAward(taskType, taskTitle);
+                BaseModel.sleepTaskInterval();
             }
         }
         catch (Throwable t) {